refactor: RockPaperScissrosGame 구조체로 변경 및 파일 분리
